requirements.txt
asgiref==3.10.0
attrs==25.4.0
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
drf-spectacular==0.29.0
fonttools==4.60.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
uritemplate==4.2.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
